[BUGFIX] Fix a single sample seen as missing:
For the first sample the `samples_api` will return 0 as this is the first index. Checking `if not 0` will evaluate to True and append the first sample to the list of missing samples, which causes the bug.
The condition should check for None and as discussed here:
https://stackoverflow.com/questions/14247373/python-none-comparison-should-i-use-is-or
the `is` operator should be used in this case.

TODO: As coming up with a condition where a single result is in the database but is not returned by the API is tricky, create a fake dataset to use as a test for the comparing procedure.