Revert the HA parameter and add IP dependency:
Apparently our policies deny the user to specify whether their want a HA
router or not, so revert the template to where it was before.

Also, Heat may want to assign the floating IP before attaching a private
subnet to the router which will also cause the stack creation to fail.